Updates to get cards showing

+++ MagicTheGatheringArenaDeckMaster/MagicTheGatheringArena.Core/Services/ScryfallService.cs
@@ -89,21 +89,25 @@ namespace MagicTheGatheringArena.Core.Services
 
                     //string name = uniqueDataType.image_uris.small.Substring(index, count).Replace(".jpg", "");
                     string name = uniqueDataType.name_field.ReplaceBadWindowsCharacters();
-                    string fullPath = Path.Combine(setPath, name);
                     string fullName = name + "-small.jpg";
+                    string fullPath = Path.Combine(setPath, name);
                     string absolutePathToFile = Path.Combine(fullPath, fullName);
 
-                    HttpClient client = new HttpClient();
+                    // only download the image if we don't have it
+                    if (!File.Exists(absolutePathToFile))
+                    {
+                        HttpClient client = new HttpClient();
 
-                    HttpResponseMessage response = client.GetAsync(uniqueDataType.image_uris.small, HttpCompletionOption.ResponseHeadersRead).Result;
-                    response.EnsureSuccessStatusCode();
+                        HttpResponseMessage response = client.GetAsync(uniqueDataType.image_uris.small, HttpCompletionOption.ResponseHeadersRead).Result;
+                        response.EnsureSuccessStatusCode();
 
-                    byte[] imageBytes = response.Content.ReadAsByteArrayAsync().Result;
+                        byte[] imageBytes = response.Content.ReadAsByteArrayAsync().Result;
 
-                    // make sure our directory exists
-                    if (!Directory.Exists(fullPath)) Directory.CreateDirectory(fullPath);
+                        // make sure our directory exists
+                        if (!Directory.Exists(fullPath)) Directory.CreateDirectory(fullPath);
 
-                    File.WriteAllBytes(absolutePathToFile, imageBytes);
+                        File.WriteAllBytes(absolutePathToFile, imageBytes);
+                    }
 
                     ImageProcessed?.Invoke(this, 1); // we processed one image so just send 1 along
                 }
@@ -114,22 +118,27 @@ namespace MagicTheGatheringArena.Core.Services
                     int index2 = uniqueDataType.image_uris.normal.LastIndexOf("?", StringComparison.OrdinalIgnoreCase);
                     int count = index2 - index;
 
-                    string name = uniqueDataType.image_uris.normal.Substring(index, count).Replace(".jpg", "");
+                    //string name = uniqueDataType.image_uris.normal.Substring(index, count).Replace(".jpg", "");
+                    string name = uniqueDataType.name_field.ReplaceBadWindowsCharacters();
                     string fullName = name + "-normal.jpg";
-                    string fullPath = Path.Combine(pathing.CardImagePath, name);
+                    string fullPath = Path.Combine(setPath, name);
                     string absolutePathToFile = Path.Combine(fullPath, fullName);
 
-                    HttpClient client = new HttpClient();
+                    // only download the image if we don't have it
+                    if (!File.Exists(absolutePathToFile))
+                    {
+                        HttpClient client = new HttpClient();
 
-                    HttpResponseMessage response = client.GetAsync(uniqueDataType.image_uris.normal, HttpCompletionOption.ResponseHeadersRead).Result;
-                    response.EnsureSuccessStatusCode();
+                        HttpResponseMessage response = client.GetAsync(uniqueDataType.image_uris.normal, HttpCompletionOption.ResponseHeadersRead).Result;
+                        response.EnsureSuccessStatusCode();
 
-                    byte[] imageBytes = response.Content.ReadAsByteArrayAsync().Result;
+                        byte[] imageBytes = response.Content.ReadAsByteArrayAsync().Result;
 
-                    // make sure our directory exists
-                    if (!Directory.Exists(fullPath)) Directory.CreateDirectory(fullPath);
+                        // make sure our directory exists
+                        if (!Directory.Exists(fullPath)) Directory.CreateDirectory(fullPath);
 
-                    File.WriteAllBytes(absolutePathToFile, imageBytes);
+                        File.WriteAllBytes(absolutePathToFile, imageBytes);
+                    }
 
                     ImageProcessed?.Invoke(this, 1); // we processed one image so just send 1 along
                 }
@@ -140,22 +149,26 @@ namespace MagicTheGatheringArena.Core.Services
                     int index2 = uniqueDataType.image_uris.large.LastIndexOf("?", StringComparison.OrdinalIgnoreCase);
                     int count = index2 - index;
 
-                    string name = uniqueDataType.image_uris.large.Substring(index, count).Replace(".jpg", "");
+                    string name = uniqueDataType.name_field.ReplaceBadWindowsCharacters();
                     string fullName = name + "-large.jpg";
-                    string fullPath = Path.Combine(pathing.CardImagePath, name);
+                    string fullPath = Path.Combine(setPath, name);
                     string absolutePathToFile = Path.Combine(fullPath, fullName);
 
-                    HttpClient client = new HttpClient();
+                    // only download the image if we don't have it
+                    if (!File.Exists(absolutePathToFile))
+                    {
+                        HttpClient client = new HttpClient();
 
-                    HttpResponseMessage response = client.GetAsync(uniqueDataType.image_uris.large, HttpCompletionOption.ResponseHeadersRead).Result;
-                    response.EnsureSuccessStatusCode();
+                        HttpResponseMessage response = client.GetAsync(uniqueDataType.image_uris.large, HttpCompletionOption.ResponseHeadersRead).Result;
+                        response.EnsureSuccessStatusCode();
 
-                    byte[] imageBytes = response.Content.ReadAsByteArrayAsync().Result;
+                        byte[] imageBytes = response.Content.ReadAsByteArrayAsync().Result;
 
-                    // make sure our directory exists
-                    if (!Directory.Exists(fullPath)) Directory.CreateDirectory(fullPath);
+                        // make sure our directory exists
+                        if (!Directory.Exists(fullPath)) Directory.CreateDirectory(fullPath);
 
-                    File.WriteAllBytes(absolutePathToFile, imageBytes);
+                        File.WriteAllBytes(absolutePathToFile, imageBytes);
+                    }
 
                     ImageProcessed?.Invoke(this, 1); // we processed one image so just send 1 along
                 }
@@ -166,22 +179,26 @@ namespace MagicTheGatheringArena.Core.Services
                     int index2 = uniqueDataType.image_uris.png.LastIndexOf("?", StringComparison.OrdinalIgnoreCase);
                     int count = index2 - index;
 
-                    string name = uniqueDataType.image_uris.png.Substring(index, count).Replace(".png", "");
+                    string name = uniqueDataType.name_field.ReplaceBadWindowsCharacters();
                     string fullName = name + "-PNG.png";
-                    string fullPath = Path.Combine(pathing.CardImagePath, name);
+                    string fullPath = Path.Combine(setPath, name);
                     string absolutePathToFile = Path.Combine(fullPath, fullName);
 
-                    HttpClient client = new HttpClient();
+                    // only download the image if we don't have it
+                    if (!File.Exists(absolutePathToFile))
+                    {
+                        HttpClient client = new HttpClient();
 
-                    HttpResponseMessage response = client.GetAsync(uniqueDataType.image_uris.png, HttpCompletionOption.ResponseHeadersRead).Result;
-                    response.EnsureSuccessStatusCode();
+                        HttpResponseMessage response = client.GetAsync(uniqueDataType.image_uris.png, HttpCompletionOption.ResponseHeadersRead).Result;
+                        response.EnsureSuccessStatusCode();
 
-                    byte[] imageBytes = response.Content.ReadAsByteArrayAsync().Result;
+                        byte[] imageBytes = response.Content.ReadAsByteArrayAsync().Result;
 
-                    // make sure our directory exists
-                    if (!Directory.Exists(fullPath)) Directory.CreateDirectory(fullPath);
+                        // make sure our directory exists
+                        if (!Directory.Exists(fullPath)) Directory.CreateDirectory(fullPath);
 
-                    File.WriteAllBytes(absolutePathToFile, imageBytes);
+                        File.WriteAllBytes(absolutePathToFile, imageBytes);
+                    }
 
                     ImageProcessed?.Invoke(this, 1); // we processed one image so just send 1 along
                 }
@@ -192,22 +209,26 @@ namespace MagicTheGatheringArena.Core.Services
                     int index2 = uniqueDataType.image_uris.art_crop.LastIndexOf("?", StringComparison.OrdinalIgnoreCase);
                     int count = index2 - index;
 
-                    string name = uniqueDataType.image_uris.art_crop.Substring(index, count).Replace(".jpg", "");
+                    string name = uniqueDataType.name_field.ReplaceBadWindowsCharacters();
                     string fullName = name + "-artCrop.jpg";
-                    string fullPath = Path.Combine(pathing.CardImagePath, name);
+                    string fullPath = Path.Combine(setPath, name);
                     string absolutePathToFile = Path.Combine(fullPath, fullName);
 
-                    HttpClient client = new HttpClient();
+                    // only download the image if we don't have it
+                    if (!File.Exists(absolutePathToFile))
+                    {
+                        HttpClient client = new HttpClient();
 
-                    HttpResponseMessage response = client.GetAsync(uniqueDataType.image_uris.art_crop, HttpCompletionOption.ResponseHeadersRead).Result;
-                    response.EnsureSuccessStatusCode();
+                        HttpResponseMessage response = client.GetAsync(uniqueDataType.image_uris.art_crop, HttpCompletionOption.ResponseHeadersRead).Result;
+                        response.EnsureSuccessStatusCode();
 
-                    byte[] imageBytes = response.Content.ReadAsByteArrayAsync().Result;
+                        byte[] imageBytes = response.Content.ReadAsByteArrayAsync().Result;
 
-                    // make sure our directory exists
-                    if (!Directory.Exists(fullPath)) Directory.CreateDirectory(fullPath);
+                        // make sure our directory exists
+                        if (!Directory.Exists(fullPath)) Directory.CreateDirectory(fullPath);
 
-                    File.WriteAllBytes(absolutePathToFile, imageBytes);
+                        File.WriteAllBytes(absolutePathToFile, imageBytes);
+                    }
 
                     ImageProcessed?.Invoke(this, 1); // we processed one image so just send 1 along
                 }
@@ -218,22 +239,26 @@ namespace MagicTheGatheringArena.Core.Services
                     int index2 = uniqueDataType.image_uris.border_crop.LastIndexOf("?", StringComparison.OrdinalIgnoreCase);
                     int count = index2 - index;
 
-                    string name = uniqueDataType.image_uris.border_crop.Substring(index, count).Replace(".jpg", "");
+                    string name = uniqueDataType.name_field.ReplaceBadWindowsCharacters();
                     string fullName = name + "-borderCrop.jpg";
-                    string fullPath = Path.Combine(pathing.CardImagePath, name);
+                    string fullPath = Path.Combine(setPath, name);
                     string absolutePathToFile = Path.Combine(fullPath, fullName);
 
-                    HttpClient client = new HttpClient();
+                    // only download the image if we don't have it
+                    if (!File.Exists(absolutePathToFile))
+                    {
+                        HttpClient client = new HttpClient();
 
-                    HttpResponseMessage response = client.GetAsync(uniqueDataType.image_uris.border_crop, HttpCompletionOption.ResponseHeadersRead).Result;
-                    response.EnsureSuccessStatusCode();
+                        HttpResponseMessage response = client.GetAsync(uniqueDataType.image_uris.border_crop, HttpCompletionOption.ResponseHeadersRead).Result;
+                        response.EnsureSuccessStatusCode();
 
-                    byte[] imageBytes = response.Content.ReadAsByteArrayAsync().Result;
+                        byte[] imageBytes = response.Content.ReadAsByteArrayAsync().Result;
 
-                    // make sure our directory exists
-                    if (!Directory.Exists(fullPath)) Directory.CreateDirectory(fullPath);
+                        // make sure our directory exists
+                        if (!Directory.Exists(fullPath)) Directory.CreateDirectory(fullPath);
 
-                    File.WriteAllBytes(absolutePathToFile, imageBytes);
+                        File.WriteAllBytes(absolutePathToFile, imageBytes);
+                    }
 
                     ImageProcessed?.Invoke(this, 1); // we processed one image so just send 1 along
                 }
